•	Freeway Traffic around a Bascule Bridge in 2D Using OpenGL
Institution: HKBK College of Engineering, Bangalore
Abstract: In this project I have incorporated and presented the working
environment of the Freeway Traffic around a Suspension Bridge also known
as “Bascule Bridge in 2D” using OpenGL. The bascule bridge works with a
counterweight that balances the span (leaf) while the upward swing
provides clearance for boat traffic. Here, we have created a scene
consisting of the Bascule Bridge which operates to allow a Boat to pass
under it and an Aeroplane passing over it while a Bus waits for the leaf
of the Bridge to swing back into its original position and then passes
along the Bridge after the Boat and Aeroplane have sailed across. Thus,
all the 3 modes of the traffic-Air, Water and Road are handled. All the
components have been clearly demarcated to cater even to a novice user.
An extensive user-interface system is provided through a popup menu with
all the toggles attached to the right mouse button. We have even
provided keyboard interface to start, stop motion and to exit the
window.